Support non-zero starting offset in client test
If a newly created partition contains an initial control batch then it
will appear that the new "empty" partition starts at offset > 0. This
patch updates the test to not make an assumption that the starting
offset is 0 and instead initializes the starting offset based on the
first observed response from fetch/produce.

Signed-off-by: Noah Watkins <noah@vectorized.io>