Adjust for control batches when fetching
Effectively what we do here is instead of fetching and comparing with
the expected set of messages we compare the suffixes. The idea is that
we assume the suffixes are the same but the offsets are different if the
prefixes aren't the same so adjust for that. They may differ in redpanda
because the log starts off with a control batch.

Signed-off-by: Noah Watkins <noah@vectorized.io>